Crime_Records_Analysis_Final_Fixed_Connection/requirements.txt
pandas
numpy
mysql-connector-python
matplotlib
